% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-placer.R
\name{angle_halfcircle_bottom}
\alias{angle_halfcircle_bottom}
\alias{angle_halfcircle_right}
\alias{angle_zero}
\alias{angle_identity}
\title{Standardize label angles}
\usage{
angle_halfcircle_bottom(theta)

angle_halfcircle_right(theta)

angle_zero(theta)

angle_identity(theta)
}
\arguments{
\item{theta}{Numeric vector of angles in radians.}
}
\description{
Functions to standardize rotation angles to specific angle ranges.
}
\details{
\code{angle_halfcircle_bottom()} standardizes angles to (-pi/2, pi/2].
\code{angle_halfcircle_right()} standardizes angles to (0, pi].
\code{angle_halfcircle_zero()} sets all angles to 0.
\code{angle_identity()} does not modify any angles.
}
