% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{clip_lines_impl}
\alias{clip_lines_impl}
\title{Clip lines to the outside of a box}
\usage{
clip_lines_impl(x, y, id, p_mid_x, p_mid_y, width, height, theta)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{id}{Integer vector of id numbers indicating which lines are connected}

\item{width}{Box width}

\item{height}{Box height}

\item{theta}{Box angle, in radians}

\item{p_mid}{Numeric vector specifying x and y position of the box midpoint}
}
\description{
Clip lines to the outside of a box. The box is specified via midpoint, width,
height, and a rotation angle in radians. This is used to create space within
isolines for text labels or other annotations.
}
