% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw-key.R, R/geom-isobands.R
\docType{data}
\name{draw_key_polypath}
\alias{draw_key_polypath}
\alias{geom_isobands}
\alias{GeomIsobands}
\title{Draw isoband and isoline contours}
\usage{
geom_isobands(mapping = NULL, data = NULL, stat = "isolevels",
  position = "identity", ..., na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE)
}
\description{
Draw isoband and isoline contours
}
\examples{
ggplot(mapping = aes(eruptions, waiting)) +
  geom_point(data = faithful, size = 0.3) +
  geom_isobands(
    data = faithfuld,
    aes(z = density, fill = stat(level)),
    n = 10, alpha = 0.5, size = 0.2
  ) +
  scale_fill_viridis_c(guide = "legend")

volcano3d <- reshape2::melt(volcano)
names(volcano3d) <- c("x", "y", "z")

ggplot(volcano3d, aes(x, y, z = z)) +
  geom_isobands(aes(color = stat(zmin)), fill = NA) +
  scale_color_viridis_c() +
  coord_cartesian(expand = FALSE) +
  theme_bw()

ggplot(volcano3d, aes(x, y, z = z)) +
  geom_isobands(aes(fill = stat(zmin)), fill_alpha = 0.5, size = 0.2) +
  scale_fill_viridis_c(guide = "legend") +
  coord_cartesian(expand = FALSE) +
  theme_bw()
}
\keyword{datasets}
