% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoband-package.R
\docType{package}
\name{isoband-package}
\alias{isoband}
\alias{isoband-package}
\title{isoband: Generate Isolines and Isobands from Regularly Spaced Elevation Grids}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A fast C++ implementation to generate contour lines (isolines) and contour polygons (isobands) from regularly spaced grids containing elevation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://isoband.r-lib.org}
  \item \url{https://github.com/r-lib/isoband}
  \item Report bugs at \url{https://github.com/r-lib/isoband/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})
  \item Claus O. Wilke \email{wilke@austin.utexas.edu} (\href{https://orcid.org/0000-0002-7470-9261}{ORCID}) (Original author)
}

Other contributors:
\itemize{
  \item Posit, PBC (\href{https://ror.org/03wc8by49}{ROR}) [copyright holder, funder]
}

}
\keyword{internal}
