% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolines-grob.R
\name{isolines_grob}
\alias{isolines_grob}
\title{Render labeled isolines}
\usage{
isolines_grob(lines, gp = gpar(), breaks = NULL, labels = NULL,
  units = "npc", label_col = NULL, label_alpha = NULL)
}
\arguments{
\item{lines}{Isolines, as produced by the \code{\link[=isolines]{isolines()}} function.}

\item{gp}{Grid graphical parameters. Parameters applying to lines
(such as \code{col}, \code{lwd}, \code{lty}, etc.) are recycled among the total
number of lines drawn. Parameters applying only to labels (such
as \code{fontfamily}, \code{fontsize}) are recycled among the specified
breaks only. The two parameters \code{col} and \code{alpha} are also applied
to labels, unless overridden (see \code{label_col} and \code{label_alpha}),
but are matched to the corresponding lines.}

\item{breaks}{Character vector specifying the isolines that should be
labeled. If \code{NULL}, labels all isolines.}

\item{labels}{Character vector specifying the labels for each break.
If \code{NULL}, uses the breaks as labels. The number of labels provided
must match the number of breaks provided.}

\item{units}{A character string specifying the units in which to
interpret the isolines coordinates. Defaults to \code{"npc"}.}

\item{label_col}{Color applied to labels. Can be used to override the
color provided in \code{gp}, in case labels and lines should have different
colors.}

\item{label_alpha}{Alpha applied to labels. Can be used to override the
alpha value provided in \code{gp}, in case labels and lines should have
different alpha values.}
}
\description{
This function generates a grid grob that represents labeled isolines.
}
\examples{
x <- (1:ncol(volcano))/(ncol(volcano)+1)
y <- (nrow(volcano):1)/(nrow(volcano)+1)
lines <- isolines(x, y, volcano, 10*(10:18))

library(grid)
grid.newpage()

# create a colored background
grid.newpage()
grid.rect(width = .8, height = .8, gp = gpar(fill = "azure1", col = NA))
grid.rect(width = .6, height = .6, gp = gpar(fill = "cornsilk1", col = NA))
grid.rect(width = .4, height = .4, gp = gpar(fill = "cornsilk2", col = NA))
grid.rect(width = .2, height = .2, gp = gpar(fill = "cornsilk3", col = NA))

# draw labeled lines
grid.draw(isolines_grob(lines))
}
