% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label-placer.R
\name{label_placer_minmax}
\alias{label_placer_minmax}
\title{Set up a label placement strategy}
\usage{
label_placer_minmax(placement = "tb",
  rot_adjuster = angle_halfcircle_bottom(), n = 2)
}
\arguments{
\item{placement}{String consisting of any combination of the letters
"t", "r", "b", "l" indicating the placement of labels at the top,
to the right, at the bottom, to the left of the isoline.}

\item{rot_adjuster}{Function that standardizes the rotation angles of the labels.
See e.g. \code{\link[=angle_halfcircle_bottom]{angle_halfcircle_bottom()}}.}

\item{n}{Size of the point neighborhood over which the rotation angle should be
calculated.}
}
\description{
The minmax label placer places labels at the horizontal or vertical minima or maxima of
the respective isolines.
}
