% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-iso.R
\name{plot_iso}
\alias{plot_iso}
\title{Visualize a single isoband}
\usage{
plot_iso(m, vlo, vhi, fill_lo = "black", fill_mid = "gray50",
  fill_hi = "white", newpage = TRUE)
}
\arguments{
\item{m}{input matrix}

\item{vlo}{lower cutoff for isobanding}

\item{vhi}{higher cutoff for isobanding}
}
\description{
This function visualizes a single isoband calculated from a matrix. It is mainly useful
for debugging and visualizing the isobanding algorithm
}
