% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip-lines.R
\name{clip_lines}
\alias{clip_lines}
\title{Clip lines so they don't run into a set of boxes.}
\usage{
clip_lines(x, y, id, clip_boxes, asp = 1)
}
\arguments{
\item{x}{Numeric vector of x coordinates}

\item{y}{Numeric vector of y coordinates}

\item{id}{Integer vector of id numbers indicating which lines are connected}

\item{clip_boxes}{Data frame specifying the locations of boxes to clip to.
Should have five columns, named \code{x}, \code{y}, \code{width}, \code{height}, \code{theta}, which
specify the x and y positions of each box midpoint, as well as the box width,
box height, and box angle in radians. Each box is specified by one data
row.}

\item{asp}{Aspect ratio (width/height) of the target canvas. This is used to convert
widths to heights and vice versa for rotated boxes}
}
\description{
Clip lines so they don't run into a set of boxes. Useful for labeling isolines,
as it allows removal of line segments that would run into any text labels.
}
