% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoband.R
\docType{package}
\name{isoband-package}
\alias{isoband}
\alias{isoband-package}
\title{isoband: Generate Isolines and Isobands from Regularly Spaced Elevation Grids}
\description{
A fast C++ implementation to generate contour lines (isolines) and
  contour polygons (isobands) from regularly spaced grids containing elevation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wilkelab.org/isoband/}
  \item Report bugs at \url{https://github.com/wilkelab/isoband/issues}
}

}
\author{
\strong{Maintainer}: Claus O. Wilke \email{wilke@austin.utexas.edu} (\href{https://orcid.org/0000-0002-7470-9261}{ORCID})

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})
}

Other contributors:
\itemize{
  \item testthat and Catch authors (testthat C++ testing code) [contributor]
}

}
