% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{isobands}
\alias{isobands}
\alias{isolines}
\title{Efficient calculation of isolines and isobands from elevation grid}
\usage{
isobands(x, y, z, value_low, value_high)

isolines(x, y, z, value)
}
\arguments{
\item{x}{Numeric vector specifying the x locations of the grid points.}

\item{y}{Numeric vector specifying the y locations of the grid points.}

\item{z}{Numeric matrix specifying the elevation values for each grid point.}

\item{value_low, value_high}{Numeric vectors of minimum/maximum z values for each to be generated isoband.}

\item{value}{Numeric vector of z values for which isolines should be generated.}
}
\description{
Efficient calculation of isolines and isobands from elevation grid
}
