% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolines-grob.R
\name{isolines_grob}
\alias{isolines_grob}
\title{Render labeled isolines}
\usage{
isolines_grob(lines, gp = gpar())
}
\arguments{
\item{lines}{Isolines, as produced by the \code{\link[=isolines]{isolines()}} function.}

\item{gp}{Grid graphical parameters.}
}
\description{
This function generates a grid grob that represents labeled isolines.
}
\examples{
x <- (1:ncol(volcano))/(ncol(volcano)+1)
y <- (nrow(volcano):1)/(nrow(volcano)+1)
lines <- isolines(x, y, volcano, 10*(10:18))

library(grid)
grid.newpage()

# make some colored background
grid.rect(width = .7, height = .7, gp = gpar(fill = "cornsilk1", col = NA))
grid.rect(width = .5, height = .5, gp = gpar(fill = "cornsilk2", col = NA))
grid.rect(width = .3, height = .3, gp = gpar(fill = "cornsilk3", col = NA))

# draw labeled lines
grid.draw(isolines_grob(lines))
}
