% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isolines-grob.R
\name{isolines_grob}
\alias{isolines_grob}
\title{Render labeled isolines}
\usage{
isolines_grob(lines, gp = gpar())
}
\arguments{
\item{lines}{Isolines, as produced by the \code{\link[=isolines]{isolines()}} function.}

\item{gp}{Grid graphical parameters.}
}
\description{
This function generates a grid grob that represents labeled isolines.
}
\examples{
m <- matrix(c(0, 0, 0, 0, 0, 0,
              0, 0, 0, 1, 1, 0,
              0, 0, 1, 2, 1, 0,
              0, 1, 2, 0, 0, 0,
              0, 0, 0, 3, 0, 0,
              0, 0, 0, 0, 0, 0), 6, 6, byrow = TRUE)

lines <- isolines(
  (1:ncol(m))/(ncol(m)+1), (nrow(m):1)/(nrow(m)+1),
  m, c(0.5, 1.5, 2.5, 3.5)
)

library(grid)
grid.newpage()
grid.draw(isolines_grob(lines))
}
